requirements.txt
pydantic
fastapi
pymongo
python-dotenv
fastapi
uvicorn[standard]